FIX:14027 async agent zombie process
The set_up_path function always appended "/usr/local/bin" to the PATH
variable and was called in the main loop. This means the environment
is growing with each iteration. Most UNIXes impose a limit on the
total size of the environment [1]. Once the limit is reached any
attempt to execute a program will fail with E2BIG error, visible
via strace. With the infinite loop this leads to a zombie process
that consumes a lot of CPU.

[1] https://man7.org/linux/man-pages/man2/execve.2.html

FIX FEED-6874

Change-Id: I924fd52e40d4908c6685932a25969e0360b3757f