Improved EventContext Typing
The eventconsole has some context-keys seemingly unique to the event
console. So I subclassed the EventContext with a ECEventContext.

Also the EventContext was moved to cmk.utils so it can be imported
without violating submodule structures.

cmk.utils.defines.service_state_name offers basically a str as
service-state. So the Literal values were removed.

Change-Id: I5421dcc4fc0e4969a8dbe3d4d036d22d6cae1af9