13929 FIX checkbox tag-group handling
In previous versions when a tag-group with only one choice (a checkbox) was
set to a folder and subsequently removed on a host in that folder, the "show hosts"
REST API call on that folder crashed with a 500 error (with the error: "tag_xxx can
not be None").

This werk fixes this behaviour. Apache still needs to be restarted so newly created
tag-groups come into effect on the REST API.

Change-Id: I69124d9ec115f9a84591e712b67f1930fcae1ff2