14067 FIX REST-API host_config: addition of validation for host labels
Host labels take the key:value format where both key and value are of
the string type. In the previous version, it was possible to specify
non valid entries such as a list of strings as the REST API did not
perform validation for the labels. This consequently led to other
problems as Checkmk attempted to process the label values expecting
a string. This werk introduces validation for the host labels in
order to prevent such instances.

Change-Id: I4c4c5a57832d3def9b058e162d9136d5c3db5c21