13805 FIX No longer show dashboards specific to one host in a host menu
Previously, the host menu would show the following entries:

- Kubernetes Cluster

- Kubernetes Deployment

- Kubernetes StatefulSet

- Kubernetes DaemonSet

- Kubernetes Namespace

- Single Linux host

- Single Windows host

These dashboards are specific to the type of the underlying host.
Therefore, the entry would lead to a broken dashboard, whenever there
was a mismatch between the host expected by the dashboard and the
host, to which the menu belongs. With this werk, the entries are no
longer shown in the menu.

Werk was implemented with Change-ids:
I5590d124091de2f8eac75f9a37ccad8709f33737
I816c35e6e84e0a8c189e10d2265758590418c678

CMK-10245

Change-Id: Id875d0dc8cd79c1c659e94ee74b052dbcfb47c22