13903 SEC Introduce additional CSRF checks
Previous to this Werk most data changing requests needed a transaction id. This
mechanism was not used in all methods and in some cases the transaction id
mechanism was not a good fit. So now a new CSRF token is introduced and checks
are implemented in various locations.

If you scripted the UI you need to parse the CSRF token from the HTML and send
it with the Post requests.

Change-Id: Ic709514c4e0f00c1aeeeede8aaf4c388007fcd7a