KUBE: refactor _collect_cpu_resources test
We use Pydantic factories instead of a handrolled function.  Moreover,
we now test in a more uniform fashion and test
_collect_memory_resources alongside.

Moreover, the test is moved to the same locations as
test_collect_workload_resources_from_api_pods and renamed accordingly.

Finally, the original test file is renamed to reflect the tests is now
contains.

CMK-9525

Change-Id: If1c0d2527b5532ba072226044f3a50eb83395dc6