13647 FIX Two-factor auth: Improve error handling of wrong protocol or address
In case a user tries to register two-factor credentials without HTTPS, the
error message was not very helpful. And also the handling access via IP address
instead of a valid host- or DNS name is improved. The new message is now more
helpful and also does not raise a crash report anymore.

Change-Id: Id0c54681806f7801f18b7927a4a95633e13157ef