Fix gh actions job
Some variables are only available in the cee and cpe version. So in a
normal build pylint finds those variables in cmk/base/config.py. But in
a cre build that deletes all unneseccary files the variables are
actually not defined, triggering a pylint failure.

The error occured during a refactoring where black moved the disable
comment to a new line.

Change-Id: I84c6a49347b45682bfd5d1309460663053ae7a63