13857 Windows agent uses perflib to obtain wmi_cpuload data
Windows agent uses now the Perflib API for service wmi_cpuload
as a more reliable and stable source of information compared
to the previously used WMI API.

If, for some reasons, the new functionality causes problems,
it is possible to return to the WMI API, using the appropriate
key in the user config. To do it, set in section  global the key
cpuload_method to 'use_wmi'.

Change-Id: Iadf03ff49d0a7094142efb335637280f4436d870