Use /omd/versions/../hooks for omd config hooks
Omd executes several hooks to determin configuration options (e.g. which port
to use for the Site-Apache). These hooks are version dependend, so omd executed
these hooks via a symlink in the site to get the hooks matching the version of
the site.

The symlinks belong to the site user in order to be able to update
the version. Since a <i>omd status</i> executes those hooks as root, it was
possible for a site user to create a malicioious hook and execute code as root.

All maintained versions (>=1.6) are subject to this vulnerability. It is likely
that also previous versions were vulnerable.

CVE will be added later here.

We thank Timo Klecker for reporting this issue!

CMK-10427

Change-Id: I3ad117773aa90b5a52ea4492ee37a66e48f3fb66