Extract parts of the HTMLGenerator
The inheritance hierarchy under htmllib is suboptimal.  The
HTMLGenerator does not actually implement the ABCHTMLGenerator but
extends a class that is already very large.

This results in several cyclic dependencies involving htmllib and,
most visibly, Breadcrumb and foldable_container.

Let us fix this by splitting the functions involved in the cycles off
the class.

CMK-10313

Change-Id: I3e45697d47f3486cb62886f6074ff9b3c4bc654b