14069 FIX KUBE agent_kube: namespace filtering and monitoring
The memory and CPU resources checks depend on the section from the Kubernetes API
and optionally on the section from the cluster collector. For some namespace hosts,
only the latter was available which led to a crash. This werk fixes this behavior
and omits the cluster collector section if the API section is not available.
Furthermore, this werk restricts hosts creation to namespaces that have at least
one running or pending pod.

Change-Id: Ie90a88b9f92f18d183902969fa0485a631bb7a26