transform: introduce json parsing for v1.23+
The Kubernetes python client v1.23 is unable to handle servers with
version v1.22 and below. For this reason we cannot upgrade the client.
However, without the upgrade, it is not possible to support features,
which have been introduced in v1.23 and above. We introduce our own JSON
parsing to cope with this issue. Intially, we only provide the parsing
for StatefulSets in order to unblock the issue in CMK-10232.

The mixing optional and required keys of TypedDicts via inheritance is
based on https://peps.python.org/pep-0589/#totality
An alternate syntax was rejected (see same PEP).

CMK-10428

Change-Id: I347dca8b9232377bebe5710f1541dbb2ed981df1