transform: remove response_type from _request
This argument was incorrecly implemented and therefore call_api would
always be called with response_type=str. Thus, the reponse_type would
only serve as a way to set the typing for response.  However, the
assumption that call_api returns the type given by response_type is
also incorrect. Example: If call_api is called with the argument
`_preload_content=False`, then a urlib3 object is returned and not
response_type.

Since this argument serves no purpose, we remove the argument.

Change-Id: I0a148184720c4faddb9f190e5ac887fa277987f2