transform: remove response_type from api_call
The argument response_type=str causes api_call to return a str, which
is consistent with the OpenAPI spec. However, this will invalidate any
json data, which we obtain from an http request. Previously, we would
use ast to deal with this issue. With the new arguments, api_call no
longer uses the auto-generated code for parsing the http response.
Instead, we implement our own parsing inside of RawAPI.

Change-Id: Ib4f2f0c44df4edcf5dac7791404e656a99f7ca30