14126 KUBE agent_kube: clarify error messages caused by unsupported Kubernetes version
With this werk, the agent verifies whether the Kubernetes version is
supported by Checkmk. If it is known that an API server with a certain
version is incompatible, the agent will now provide a helpful error
message.  Different flavors of Kubernetes are not verified. In the
event, that the version of the API server deviates from the versioning
scheme used by Kubernetes, processing the API data is still attempted
on a best effort basis. Errors related to unknown versioning schemes
can be inspected by increasing the verbosity of the agent via the
command line (you can find the command line tool by using cmk -D).

CMK-10198

Change-Id: I398cf6e05b595873ab5f376788ec5f7d3c28065f