14049 cmk-agent-ctl: Agent controller can be disabled on Linux
It was possible to disable agent controller in agent bakery for Windows.
Now, it's also possible for the Linux distributions.

Disabling the agent controller is not advised unless you experience
problems with it. In this case the agent will work in legacy pull mode.
Please note: In the legacy pull mode the transported data is not encrypted.

CMK-9895

Change-Id: I5909080f21a88750b46ec26248b42a907bc77e97