14127 FIX KUBE agent_kube: Option Cluster resource aggregation matches inline help
Previously, if the option 'Cluster resource aggregation' was used to
exclude Nodes based on their roles, the performed search would be a
prefix search. This was inconsistent with the inline help, which
promises an infix search. With this werk, an infix search is now
performed by the agent. Those users who have configured the option
'Cluster resource aggregation' to be sensitive to the type of regex
matching performed need to verify that the agent is still performing
as expected.

Change-Id: Iaf5df0e4d44af82a1f47aaf83f2746c8c0679f85