14128 FIX KUBE agent_kube: Option Monitor namespaces matches inline help
Previously, if the option 'Monitor namespace' was used to filter
namespaces, the performed search would be a prefix search. This was
inconsistent with the inline help, which promises a complete search.
With this werk, an infix search is now performed by the agent, which
is the default in Checkmk. Those users who have configured the option
'Monitor namespaces' to be sensitive to the type of regex matching
performed need to verify that the agent is still performing as
expected.

Change-Id: Ibc223a8c8498222b4fa802fe3b36fa3019b34319