14129 FIX KUBE kube_pod_status: Option Interpretation of pod status matches inline help
Previously, if the option 'Interpretation of pod status' was used to
assign a Pod to a group, the performed search would be a prefix
search. This was inconsistent with the inline help, which promises an
infix search. With this werk, an infix search is now performed by the
agent. Those users who have configured the option 'Interpretation of
pod status' to be sensitive to the type of regex matching performed
need to verify that the agent is still performing as expected.

Change-Id: I3a298e912554a11a8e1a59bdd8081330d9e27fbd