Simplified CSSSpec type alias.
Having both a list[str] and a list[Optional[str]] in a Union is not really
helpful: Python's Unions are not tagged, so we can't distinguish between
those alternatives.  Consider e.g. the case when we have a ["hurz"] at
hand: Is it a list[str] or a list[Optional[str]]?  We simply don't know.
The same holds for [], etc. etc.

With this simplification, dispatching on a CSSSpec value is easy now: Is it
None?  Is it a str?  Or something else (i.e. a list)?  The last case will
probably change to a Sequence in the future, so we have to do the dispatch
in that order, because a str is a Sequence, too.

Change-Id: I536a848caf2e4cecd8256c67ddb042f56c576e25