Simplify CSSSpec, part 1.
A single str is just a shorthand for a singleton list, nuke the former.

Allowing a ton of "convenient" shorthands is a bad design principle when a
single general way of expressing things exists.  Trading a few keystrokes
during the initial development for a ton of complexity later (like
incomprehensible types, lots of dispatching code scattered all around, etc.)
is a fundamental misconception about what is "easy" and/or "convenient".

But the real "fun stuff" is mixing various overlapping Unions into one, see
e.g.  HTMLTagAttributeValue.  This makes dispatching later basically
impossible or nonsensical (choose one!).

Change-Id: Ib9488dae67848c513a4507c018e37b1591478f12