FileCache: Move `use_only_cache` to cache
It is obviously a caching option and not a fetching
option.  Let us make that clear not only from the
name but also from the organization of the code.

Note: The option did exist in the push_agent(_fetcher)
but it was not used.  It is not clear whether this was
intended or not.

Change-Id: Ia9209be93e9b9c6ae6af20ba7ee1ff78cf44e7c2