13860 FIX Windows agent internal port set to 28250
The change is intended to avoid the situation when the internal
agent port is randomly 'stolen' breaking the monitoring.
The former port 500001 is located  in the range 32768-65535
and ports from the range  may be used(randomly) by TCP clients
or even blocked by Windows(depends from version and service pack).

Change-Id: Ib875f1b4d36d43eb71ffe3f00dbd3486190ed028