use `npm ci` instead of `npm install` in Makefile
`npm install` resolves and install dependencies and generates
a new `package-lock.json` file. This is inappropriate for a CI
environment, because specific resolutions might differ locally and
and on the CI machine. Therefore, we should use `npm ci`, as it skips
the resolving part and only installs dependencies based on the
lock file.

Change-Id: Idd7f812a811fbd92e93979b34f61552770a1e19d