13421 FIX mk_sap: detect recursion in SAP data
mk_sap.py handles a tree of SAP objects. The objects are annotated with
the path from the root of the tree to the object. The objects returned
by SAP contain a link to their parent object. SAP can be configured in a
way that the object references itself as its parent. This can lead to a
cryptic recursion error message:

<code>
Unhandled exception (maximum recursion depth exceeded)
</code>

After this werk a better error message will be shown including the
object that is the root cause of the recursion error:

<code>
Could not calculate path, recursion limit reached.
Reorganise your SAP data to get past this error.
Element that causes this:
...
</code>

You will have to fix this problem within SAP.

Change-Id: I2b550ae883e6e2df9a3dee0fb6f951b332941631