FileCache: Collapse booleans #2
Remove case where the attribute `use_outdated` could be False but
the FileCache would override the setting using a non-obviously
related attribute.

This further aims at more clearly separating configuration and
business logic.  The business logic should drive the configuration
but the FileCache instances should not need to contribute to the BL
but either read the data from a file, or not.

CMK-10578

Change-Id: I905d97c4a2b9ecceccc20acbb26b72f62ce1763a