Additional fix for pip command
Since the 3.10 change, the pip command contains the
shebang #!/usr/bin/python. The sed we have in place to replace the
shebang with #!/usr/bin/env python3 did not apply anymore.

Seems strange that pip3 now points to #!/usr/bin/python since a lot of
systems still have pointing it to Python 2, but let's deal with this
anyways.

Change-Id: Ib07947ad33bcaa5f7ac6dd60a98e93b14d8dabc3