update-pos: Improve workflow
Previously, we first stripped away code comments from the po files and
then we checked for errors concerning Python formatters. This had the
disadvantage that in case of errors, the line numbers of these errors
referred to the already stripped file, which makes fixing these errors
unnecessarily difficult.

Now, we first check for errors and immediately abort if we find any.
This means that the line numbers now correspond to the unstripped po
files, which are directly available in the checkmk-locale git.

Change-Id: Ie91e1022bd5551f0aefed1a9dd37219af0119080