run-in-docker: enable containerized and native work in the same worktree
In short this change
* .. allows both containerized and native scenario in the same worktree
* .. only addresses local development
* .. introduces a $HOME directory for the container
* .. should not affect the way scripts are being run
* .. does not affect the native scenario
* .. should not affect CI builds

Currently a given check_mk clone/worktree has to be used in
acontainerized or a native context. Both at a time are not possible
because of conflicting artifacts directories (currently only
`omd/build/` and `.venv/`). In an ideal world those directories could be
located in separate dedicated build directories but that's currently not
the case.
This change makes `run-in-docker.sh` provide separate artifact
directories for the containerized case by mapping them on top of those
for the native build. Additionally a dedicated $HOME directory is being
created and passed as environment variable to address caching and user
based state.

There is room for some improvement regarding initial build dependencies
forcing `make .venv` to be triggered every time a command is being
exectuted in containerized context (doesn't take too long, though).

Change-Id: Ib8ea77b97518c9a1a6808f44ff0e0ca0cfde8a23