drop uneffective deletions
The new inventory API does not support deleting information,
once it's in the inventory.

If the information is not correct, then it should not have been
put in the HaSI in the first place.

If this really is an issue, we must fix the plugins wrongly
providing these values.

(Also this does not make sense anyway: What if the plugin
setting these values is called after this plugin?)

Change-Id: Ib235907382fa73d3145ade5ff82dc4f51394b1fa