14342 FIX agent_prometheus & prometheus_build_info: change to display version of main_instance
Previously, the agent returned an error if the build_info PromQL returned multiple entries. This
effectively prevented the user from using the agent for a Prometheus cluster setup. This werk removes
this limitation and adjusts the Prometheus Build Service accordingly. The customization is only visible
for Prometheus instances running below version 2.14 and the details can be found in the updated man page

Change-Id: I2c4e86c59c26411c3dadf26979a1070d8d250d13