14349 FIX Fix performance bug when using a cluster
Running all checks on a cluster with more than 100 services took several minutes. This has been fixed and runtime is now a few seconds.
No user interaction required.

below is a more detailed description of the changes

Refactor of value store
=======================

Add hostname as a key to the value store. The hostname
can be set explicitly when creating a namespace. By default
it choses the hostname of the host for the valuestore.

This change is introduced because a cluster consists of
multiple nodes and we need to track entries of the valuestore
for each node in the cluster separately.

During an update we migrate the counter variables to include the hostname

Refactor cluster check
======================

- Only load value store once for a cluster
- use hostname specific namespaces when evaluating checks on a cluster

Change-Id: I3f045e840c40202a4e5fc36a7e7a08990c35f29e