13683 FIX Solaris agent update removes services symlink
When updating agents using the agent updater on Solaris, the postremove script
will try to remove the corresponding service line from <tt>/etc/services</tt>.
However, the script recently missed the fact that this usually is a symbolic
link to the actual services file, and overrode it with a new (hardlinked) file.

Moreover, this situation could lead to a missing Checkmk agent services, as the
situation got cleaned up on calls to <tt>pkg verify</tt> and <tt>pkg fix</tt>
by replacing the <tt>/etc/services</tt> file with the proper symbolic link
again, locking out the Checkmk agent.

Change-Id: I553ff3dd790214f24d7c1d0b4ffa92287883dce5