14343 FIX KUBE agent_kube: change ContainerTerminatedState's start and end time to be optional
Previously, the Kubernetes special agent raised the error "[special_kube] TypeError: Can not convert
to timestamp: 'None' of type <class 'NoneType’>" when a terminated container had no information
about either its start or finish time. With this werk, the assumption that the start and finish
times are always set is removed, and is thus consistent with the Kubernetes specification.
The check plug-in is also modified to display the time-related information depending on what
fields are returned.

FEED-6963

Change-Id: Id673d5f0e762d05b5c22f7584081a0ac4a56f8c3