REST API endpoint for CSRs: restrict to CNs which are valid UUIDs
This internal endpoint is used during the agent registration process. It
issues agent certificates which are later used to encrypt the transport
of agent data. We want to make sure that we only issue agent
certificates with CNs which are valid UUIDs. This mitigates possible
attack vectors which involve the misuse of agent certificates to act as
a false agent controller.

CMK-10708

Change-Id: I62e683c43a45bb947abaf8886a6a7365a7ed264a