REST API: fix WSGI environment leading to wrong URLs
Apache's mod_wsgi sends a SCRIPT_NAME and PATH_INFO at the same
time, which Werkzeug concatenates. SCRIPT_NAME corresponds to
the "root_path" (i.e. the location where the application is mounted
under) and PATH_INFO the path relative to that. In our case
PATH_INFO already contains all the information, so having something
in SCRIPT_NAME is redundant.

This change introduces a middleware which recreates the environment
of Apache, so that we can write unit-tests against that behavior.

Change-Id: I321eac741fa79b88b37ebf77d383b5ab8274decc