ignore passive checks in the scheduled_rate for the speedometer
The Service Speed-O-Meter is intended to show the actual_check_rate
divided by the scheduled_check_rate as a percentage.  Perfect performance
is 100%.  Slow performance is indicated by numbers below 100%.

The problem we have is that scheduled_rate is artificially inflated
by the legacy nsca passive checks.

The live status query used for each site is a sum of the
1/check_interval for each check.  Check_interval is specified in minutes.

    lq "GET services\nStats: suminv check_interval"

A check that runs every two hours (like Check_MK Discovery) has a
check_interval of 120.  The passive checks have a check_interval of 1.

This means that every passive check is creating an expectation of
1 check per minute, but that rarely happens.

This patch removes passive checks from the scheduled_rate.