13932 FIX non JSON 500 page in REST API
In some error cases, a non-JSON result page was returned by the REST
API. This led to problems with script execution as the page was not
valid JSON and thus the parsers threw errors.

The errors which led to the non-JSON result will now be returned as
a JSON result and a log entry will be made.

Specifically:

 * the permission system checks that all declared permissions
   are used and all used permissions are declared
 * whenever that was not the case an exception was raised
 * this exception was not handled by the REST API by default
 * now the exception is no longer raised, but an explicit log-
   entry is made and an explicit error result is returned

Change-Id: Ib915ee66ea0d0db8f6d5c02e6d4e535fb3267f3d