Improve base config typing
Be more specific in the default config declarations. This improves the
situation only a little bit.

Typing the ruleset values would make a much bigger difference. This will
be possible as of Python 3.11. But it seems we don't have to wait for
it since typing_extensions will make it available with the next version
(https://github.com/python/typing_extensions/blob/main/CHANGELOG.md).

Change-Id: I466b3e01f1d67c9c5a6608c114da6915b4fd2d75