13933 FIX missing Theme in REST API
Some Rulesets regarding themese and icons could crash in the REST API due
to a missing reference to a default Theme. The theme was previously omitted
in the REST API which led to this error.

This has now been fixed and a default theme is now available in the REST API
for Ruleset validation logic to access, even though it has no effect for the user.

Change-Id: I6584db78b1ff6ccb521f310e684daecda0fa0a68