rewrite gcp function tests
GCP check plugin tests have been a bit all over the place. This is the
first step to get a test concept implemented for gcp plugins.

In the tests we want to test generic behavior just once. Generic
behavior is that the check parameters are adhered to. That a plugin
behaves correct if data is missing. We do this only once because the
assumption is that all gcp check plugins use the identical
implementation function from ...agent_based.utils.gcp.

What we do want to test for every check plugin is that the configuration
of check is correct. That includes we generate the correct number of
results and metrics, metrics have the expected names, and results have
the correct summary and details. Those vary between plugins. By using
the generic test data generator that works on the SERVICE definitions of
the gcp agent we can also ensure that the check plugin configurations
look for the correct timeseries in the section data.

After this all gcp test can skip the generic behavior test without losing
any test coverage.

Change-Id: Ic2b049e660ad78904b257613fed7e251437fee58