14486 FIX 'Host groups' view: Add coloring for host states
In the view "Host groups" host state cells were not colored according to their state. This is fixed, so now whenever there's at least one occurence of the given host state for the given host group, the table cell is colored. This behavior now matches that of the service state cells again.
SUP-10140

Change-Id: I7dd97170809239eb0647193d9b79b5c7b1d998ec