Fix erasing by iterator in loop containers
To erase an element by iterator and keep iterator valid,
we must always use
`it = container.erase(it);`
See cppreference and/or Meyers.
The method `container.erase(it++)` for _looping_ container
doesn't work(verified in goldbolt).

- add TODO to the erasing in the map container: probably,
we should use unified method to erase from containers.
It is not clear why Meyers recommends to use post-++ instead
of the defined by C++ standard return value from the erase().

Change-Id: I350168c0c3f6877fb5886ba40603a0da77eecfdd