14060 FIX mssql_counters: Fix crash in parsing mssql_counters
Parsing mssql_counters was failing for timestamps of different formats.
Now, we changed the mssql.vbs agent to always provide the timestamp in
the same format, independent of the localization.

If you have a crash in mssql_counters with the message 'Time string
does not match any known pattern', update the mssql.vbs agent.

SUP-10722

Change-Id: I059033ed96f6e2376d3814bac911684d18cd2027