14146 FIX Rule matching for hosts
The search for host rules in Setup did not work as expected when matching host regex conditions.

The issue did not affect the rule matching visualization nor the actual rule matching.

When checking the matching rules for hosts the rules with regular expressions
would not get a green background although the regex should match, which indicates
the rule is not matching the current search.

Now the rule matching is fixed and the rules will get a green background if they match the search.

SUP-9852

Change-Id: If6b63e6837bd50c77d870224282c9642ccea04e9