13953 FIX Prevent a crash with pending changes and Japanese locale
The system would crash if there was a pending change and the japanese locale set.
This is no longer the case.
In Japanese locale, the space was missing between the number of changes and the Japanese word for "change".
As we can not be sure how this situation is handled in different locales,
or if other locals add more spaces to this string a refactoring was necessary.

CMK-10645

Change-Id: I3cc042964b598377e59c9ab94033cba4a503f399