14064 fileinfo: Deprecate timeofday parameter
After introduction of timeperiods, timeofday ('Only check during the
following times of the day') parameter became obsolete.
With this werk, it's removed from the UI and the existing parameters
are transformed to timeperiod parameters.

If you had a combination of timeofday and timeperiod parameters,
your parameters are migrated in the following way:
- in case of timeofday in the default timeperiod, timeofday parameter is
migrated to timeperiod and other timeperiods are removed
-in case of timeofday in the non-default timeperiod, it is removed

If this migration isn't correct for your example, please fix your fileinfo
configuration.

FEED-7134

Change-Id: Ie1aa7ca47b40a6361f4f42e1e7acb9205b506644