agent_based: Add tests for the decorators
The decorators are used in the global scope and need to
access the config.  This requires that the config is
global as well.  However, we now want to have a more local
and instantiable config.  We therefore need to change the
decorators into something else.

Here, we simply add tests in order to insure that the
functionality is not lost in a later reengineering.

CMK-10515

Change-Id: I417e3b418f42350e43bc9d6c986f69f15315d466