14285 FIX Fix frozen Microcore (Livestatus not responding) during config reloads
The reload of the Microcore core and it's helper processed could freeze when
the core had notifications pending during reload. This was caused by a deadlock
between the Notification helper and the Microcore. The Microcore was still
alive but waiting to the notification helper to terminate while the
notification helper waited for the Microcore.

From the user perspective this resulted in Livestatus not being responsive
while the cmc.log showed a message like: <tt>still X unsent events, sending
them now</tt>.

Change-Id: I8ac1a2a179e44f761c21a7a7a3e0ebb3ab4cc976